agregando autenticacion automatica parte 1

+++ src/app/shared/header/header.component.ts
@@ -1,12 +1,13 @@
 import { Inject, PLATFORM_ID, Component } from '@angular/core';
 import { MenuCategoriesComponent } from './menu-categories/menu-categories.component';
 import { CommonModule } from '@angular/common';
-import { RouterModule, Router } from '@angular/router';
+import { RouterModule, Router, NavigationEnd } from '@angular/router';
 import { CookieService } from 'ngx-cookie-service';
 import { isPlatformServer } from '@angular/common';
 import { CartService } from '../../pages/home/service/cart.service';
 import { AuthService } from '../../pages/auth/service/auth.service';
 import { ToastrService } from 'ngx-toastr';
+import { filter } from 'rxjs';
 
 declare function CurrecyChange([]): any;
 declare var $: any;
@@ -43,7 +44,7 @@ export class HeaderComponent {
         }, 50);
       }, 50);
 
-      this.listadoCarrito();
+      this.validarAuthToken();
     }
   }
 
@@ -51,35 +52,58 @@ export class HeaderComponent {
     this.currency = this.cookieService.get("currency") ? this.cookieService.get("currency") : 'COP';
   }
 
+  validarAuthToken() {
+    this.router.events.pipe(
+      filter((event): event is NavigationEnd => event instanceof NavigationEnd)
+    ).subscribe(() => {
+      let token = localStorage.getItem('token');
+
+      this.authService.validarToken(token).subscribe((response: any) => {
+        if (response) {
+          this.listadoCarrito(token);
+        } else {
+          this.cartService.clearCart();
+          this.authService.tokenSubject.next(token); // Sincroniza el token
+        }
+      });
+    });
+  }
+
   deleteCart(CART: any) {
-    this.cartService.deleteCart(CART.id).subscribe((resp: any) => {
-      this.cartService.removeCart(CART);
-      this.toastr.info("El producto" + CART.product.title + "fue eliminado del carrito", "Producto eliminado");
+    let token = localStorage.getItem('token');
+
+    this.authService.validarToken(token).subscribe((response: any) => {
+      if (response) {
+        this.cartService.deleteCart(CART.id).subscribe((resp: any) => {
+          this.cartService.removeCart(CART);
+          this.toastr.info("El producto" + CART.product.title + "fue eliminado del carrito", "Producto eliminado");
+        }, (error) => {
+          console.log(error);
+          this.toastr.error('API Response - Comuniquese con el desarrollador', error.error.message || error.message);
+        });
+      } else {
+        this.cartService.clearCart();
+        this.authService.tokenSubject.next(token); // Sincroniza el token
+      }
+    });
+  }
+
+  listadoCarrito(token: any) {
+    this.authService.tokenSubject.next(token); // Sincroniza el token
+
+    this.cartService.listCart().subscribe((resp: any) => {
+      if (resp.carts.data.length > 0) {
+        resp.carts.data.forEach((cart: any) => {
+          this.cartService.changeCart(cart);
+        });
+      } else {
+        this.cartService.clearCart();
+      }
     }, (error) => {
       console.log(error);
       this.toastr.error('API Response - Comuniquese con el desarrollador', error.error.message || error.message);
     });
-  }
 
-  listadoCarrito() {
-    const token = localStorage.getItem('token');
-
-    if (token) {
-      this.authService.tokenSubject.next(token); // Sincroniza el token
-
-      this.cartService.listCart().subscribe((resp: any) => {
-        if (resp.carts.data.length > 0) {
-          resp.carts.data.forEach((cart: any) => {
-            this.cartService.changeCart(cart);
-          });
-        } else {
-          this.cartService.clearCart();
-        }
-      }, (error) => {
-        console.log(error);
-        this.toastr.error('API Response - Comuniquese con el desarrollador', error.error.message || error.message);
-      });
-    }
 
     this.cartService.currentDataCart$.subscribe((resp: any) => {
       this.listCart = resp;
